fixup! test: move the confirm mock code to its own file
