+++ packages/wrangler/src/__tests__/jest.setup.ts
@@ -1,8 +1,13 @@
-import { confirm } from "../dialogs";
+import { confirm, prompt } from "../dialogs";
 
 jest.mock("../dialogs");
 
 // By default (if not configured by mockConfirm()) calls to `confirm()` should throw.
 (confirm as jest.Mock).mockImplementation(() => {
   throw new Error("Unexpected call to `confirm()`.");
+});
+
+// By default (if not configured by mockPrompt()) calls to `prompt()` should throw.
+(prompt as jest.Mock).mockImplementation(() => {
+  throw new Error("Unexpected call to `prompt()`.");
 });
